+++ runtime/compiler/control/JITaaSCompilationThread.hpp
@@ -25,7 +25,7 @@
 
 #include <unordered_map>
 #include "control/CompilationThread.hpp"
-#include "rpc/J9Client.hpp"
+#include "net/ClientStream.hpp"
 #include "env/PersistentCollections.hpp"
 #include "env/j9methodServer.hpp"
 #include "env/J9CPU.hpp"
