+++ runtime/compiler/net/ProtobufTypeConvert.cpp
@@ -20,7 +20,7 @@
  * SPDX-License-Identifier: EPL-2.0 OR Apache-2.0 OR GPL-2.0 WITH Classpath-exception-2.0 OR LicenseRef-GPL-2.0 WITH Assembly-exception
  *******************************************************************************/
 
-#include "rpc/ProtobufTypeConvert.hpp"
+#include "net/ProtobufTypeConvert.hpp"
 
 namespace JITServer
    {
