+++ runtime/compiler/runtime/CompileService.hpp
@@ -23,7 +23,7 @@
 #ifndef COMPILE_SERVICE_H
 #define COMPILE_SERVICE_H
 
-#include "rpc/J9Server.hpp"
+#include "net/ServerStream.hpp"
 #include "j9.h"
 #include "j9nonbuilder.h"
 #include "vmaccess.h"
